update on comments
